+++ src/main/java/br/com/teste/emprestimo/payload/res/PessoaDto.java
@@ -1,12 +1,7 @@
 package br.com.teste.emprestimo.payload.res;
 
-import jakarta.persistence.Column;
-import jakarta.validation.constraints.Max;
-import jakarta.validation.constraints.NotBlank;
-
-import java.math.BigDecimal;
 import java.time.LocalDate;
 
 public record PessoaDto(Long id, String nome, LocalDate dataNascimento, String identificador, String tipoIdentificador,
-                        BigDecimal valorMinMensal, BigDecimal valorMaxEmprestimo) {
+                        Double valorMinMensal, Double valorMaxEmprestimo) {
 }
